Cleaning up test with a base class
