+++ kafka/src/test/java/io/github/atomfinger/javazone/bookstore/kafka/acceptance_test/AddBookListenerTest.given_that_we_add_a_new_book_then_new_book_should_be_added_to_db.approved.json
